chore: Use uint64

+++ progress.go
@@ -74,7 +74,7 @@ func (progress *Progress) report(rec *Recorder) {
 	progress.prevDPLen = dpLen
 	qps := float64(time.Duration(delta) * time.Second / InterimReportIntvl)
 	elapsed := time.Since(rec.StartedAt)
-	running := uint64(rec.Nagents) - progress.nDeadAgents.Load()
+	running := rec.Nagents - progress.nDeadAgents.Load()
 	width, _, err := term.GetSize(0)
 
 	if err != nil {
